requirements.txt
Flask
numpy
gunicorn
sklearn
pandas
PyYAML
dill 
matplotlib
PyYAML
evidently
requests
dill
-e .